requirements.txt
streamlit
pandas
numpy
matplotlib
plotly
geopandas
shapely
openpyxl
seaborn
Pillow